feat(ai): auto-summarize after transcribe + speaker attribution
Two related additions, both driven by the meets/ERPnext integration
that posts finished meetings via the new POST /v1/recordings endpoint
and waits for analysis to come back via webhook:

(1) Auto-summary chained after transcribe. The summary previously
required a manual button click in the dashboard, which the
server-to-server flow has no way to trigger. `transcribeRecording`
now fires `summarizeRecording` immediately after the transcript
write succeeds — fire-and-forget for the same reason as the
upload-side auto-transcribe (long meetings exceed any HTTP timeout,
and the webhook is the real completion channel).

Required extracting the summary logic out of the route handler into
`src/lib/ai/summarize-recording.ts`. The route now just authenticates,
calls the lib function, and maps the discriminated error codes onto
HTTP status. Behavior is preserved: tombstone re-check inside a
FOR UPDATE tx, the same JSON-shape parsing, the same whisper→chat
model fallback, the same transcript truncation at 8000 chars.

Two new webhook events for the chain: `summary.created` on success,
`summary.failed` on error. Added to the WEBHOOK_EVENTS allowlist;
existing endpoints that didn't subscribe to them are unaffected.

(2) Speaker attribution in the Meeting Notes preset. Local Whisper
on a CPU stack doesn't run pyannote diarization, so the model has
to infer speakers from contextual cues (self-introductions, names
addressed in dialogue, topic ownership, turn-taking). The new
prompt instructs it explicitly to: (a) use real names when clearly
attributable, (b) fall back to generic labels ("Speaker A",
"Sprecher A") in the transcript's own language, (c) omit attribution
when speakers can't be reliably separated, rather than guess. The
summary string opens with a one-line `Participants:` / `Teilnehmer:`
header so the downstream system has a stable handle.

Imperfect substitute for true diarization, but for a 2–6-person
meeting it produces drastically better output than the previous
"wall of attributed-to-nobody text".
